Transform arbitrary matrices into ARMA coefficient matrices

Creates coefficient matrices for which the characteristic polynomial corresponds to a stationary process. See Ansley and Kohn (1986) for details about the transformation used.

CoeffARMA(A, variance = NULL, ar = 1, ma = 0)

Arguments

A

An array of arbitrary square matrices in the multivariate case, or a vector of arbitrary numbers in the univariate case.

variance

A variance - covariance matrix. Note: variance not needed for the univariate case!

ar

The order of the AR part.

ma

The order of the MA part.

Value

If multivariate, a list containing:

An array of coefficient matrices for the AR part.

An array of coefficient matrices for the MA part.

If univariate, a list containing:

A vector of coefficients for the AR part.

A vector of coefficients for the MA part.

References

Ansley CF, Kohn R (1986). “A note on reparameterizing a vector autoregressive moving average model to enforce stationarity.” Journal of Statistical Computation and Simulation, 24(2), 99--106.

Author

Dylan Beijers, dylanbeijers@gmail.com

Examples

CoeffARMA(A = stats::rnorm(2), ar = 1, ma = 1)
#> $ar
#> [1] -0.6270388
#> 
#> $ma
#> [1] -0.8280127
#>